Enum cwe_checker_lib::abstract_domain::DataDomain

pub enum DataDomain<T: RegisterDomain> {
    Top(ByteSize),
    Pointer(PointerDomain<T>),
    Value(T),
}

An abstract domain representing either a pointer or a non-pointer value. Both non-pointer values and offsets of pointers are represented by the same abstract domain T.

Variants

Top(ByteSize)

The Top element of the domain. Describes a value for which nothing is known.

Pointer(PointerDomain<T>)

The value is a pointer to an abstract memory object.

Value(T)

The value is a non-pointer value or a pointer to global memory. The latter can happen if pointers to global memory are described by their absolute value.

Implementations

impl<T: RegisterDomain> DataDomain<T>

pub fn replace_abstract_id(
&mut self,
old_id: &AbstractIdentifier,
new_id: &AbstractIdentifier,
offset_adjustment: &T
)

For pointer values replace an abstract identifier with another one and add the offset_adjustment to the pointer offset. This is needed to adjust stack pointer on call and return instructions.

pub fn referenced_ids(&self) -> BTreeSet<AbstractIdentifier>

Return a set of all referenced abstract IDs. The set is empty if self is not a pointer.

pub fn remove_ids(&mut self, ids_to_remove: &BTreeSet<AbstractIdentifier>)

If self is a pointer, remove all provided IDs from the target list of it. If this would leave the pointer without any targets, replace it with Top.

impl<T: RegisterDomain + Display> DataDomain<T>

pub fn to_json_compact(&self) -> Value

Get a more compact json-representation of the data domain. Intended for pretty printing, not useable for serialization/deserialization.