Struct cwe_checker_lib::abstract_domain::DataDomain

pub struct DataDomain<T: RegisterDomain> { /* private fields */ }

An abstract domain representing a set of base values plus offsets or an absolute value (or both).

The base values are represented as abstract IDs, i.e. they are treated as variables with unknown absolute value, e.g. the returned pointer by malloc. For each base value the offset is given by an abstract domain T, which should specialize in representing absolute values (e.g. an interval domain). Note that the domain assumes pointer semantics for these values. That means if one applies operations to the domain that are not used in pointer arithmetics, the abstract ID of the base value might be removed from the domain.

If the domain also represents absolute values, then the values are given by a single instance of the abstract domain T.

The domain also contains a flag to indicate that it includes Top values, i.e. values of fully unknown origin and offset.

Implementations

impl<T: RegisterDomain> DataDomain<T>

pub fn add_offset(&self, offset: &T) -> Self

Add offset to all contained absolute and relative values of self and return the result.

pub fn subtract_offset(&self, offset: &T) -> Self

Subtract offset from all contained absolute and relative values of self and return the result.

impl<T: RegisterDomain> DataDomain<T>

pub fn is_empty(&self) -> bool

Returns true if the domain does not represent any value.

The meaning of an empty value depends on the usage of the domain. E.g. it may indicate an impossible runtime state of a program in one analysis or simply no value of interest in another analysis.

An empty value represents the bottom value in the partial order of the domain.

pub fn new_empty(size: ByteSize) -> Self

Return a new empty value with the given bytesize.

pub fn replace_abstract_id( &mut self, old_id: &AbstractIdentifier, new_id: &AbstractIdentifier, offset_adjustment: &T )

For pointer values replace an abstract identifier with another one and add the offset_adjustment to the pointer offset. This is needed to adjust stack pointer on call and return instructions.

pub fn replace_all_ids( &mut self, replacement_map: &BTreeMap<AbstractIdentifier, Self> )

Replace all abstract IDs in self with the corresponding values given by the replacement_map.

For IDs without a replacement value the contains_top_values flag will be set.

pub fn referenced_ids(&self) -> impl Iterator<Item = &AbstractIdentifier>

Return an iterator over all referenced abstract IDs.

pub fn get_relative_values(&self) -> &BTreeMap<AbstractIdentifier, T>

Return the relative values contained in the domain.

pub fn set_relative_values( &mut self, relative_values: BTreeMap<AbstractIdentifier, T> )

Replace the map of relative values with the given one.

pub fn get_absolute_value(&self) -> Option<&T>

Return the absolute value contained in the domain if present

pub fn set_absolute_value(&mut self, value: Option<T>)

Replace the absolute value contained in the domain with the given one. A value of None means that the domain does not contain an absolute value.

pub fn contains_top(&self) -> bool

Returns true if the domain contains Top values, i.e. values for which neither a value nor an abstract identifier is known.

Note that the DataDomain itself has no maximal value, i.e. this does not indicate a Top value of the abstract domain.

pub fn set_contains_top_flag(&mut self)

Indicate that the domain may contain Top values in addition to the contained absolute and relative values.

This does not remove absolute or relative value information from the domain.

pub fn unset_contains_top_flag(&mut self)

Indicate that the domain does not contain any Top values in addition to the contained absolute and relative values.

pub fn from_target(id: AbstractIdentifier, offset: T) -> Self

Return a new value representing a variable plus an offset, where the variable is represented by the given abstract ID.

pub fn remove_ids(&mut self, ids_to_remove: &BTreeSet<AbstractIdentifier>)

Remove all provided IDs from the list of relative values.

pub fn get_if_absolute_value(&self) -> Option<&T>

Return the contained absolute value only if self contains no other (relative or Top) values.

pub fn get_if_unique_target(&self) -> Option<(&AbstractIdentifier, &T)>

Return the target ID and offset of the contained relative value if self contains exactly one relative value and no absolute or Top values.

impl<T: RegisterDomain + Display> DataDomain<T>

pub fn to_json_compact(&self) -> Value

Get a more compact json-representation of the data domain. Intended for pretty printing, not useable for serialization/deserialization.