Struct cwe_checker_lib::abstract_domain::IntervalDomain

pub struct IntervalDomain { /* private fields */ }

An abstract domain representing values in an interval range with strides and widening hints.

The interval bounds are signed integers, i.e. the domain looses precision if tasked to represent large unsigned integers. The interval has a stride, i.e. all values represented by the interval are contained in the same residue class modulo the stride as the interval bounds.

The domain also contains widening hints to faciliate fast and exact widening for simple loop counter variables. See the IntervalDomain::signed_merge_and_widen method for details on the widening strategy. Note that the widening hints may not respect the stride, i.e. they may be contained in different residue classes than the interval bounds.

Implementations

impl IntervalDomain

pub fn add(&self, rhs: &Self) -> Self

Compute the interval of possible results if one adds a value from self to a value from rhs.

pub fn sub(&self, rhs: &Self) -> Self

Compute the interval of possible results if one subtracts a value in rhs from a value in self.

pub fn signed_mul(&self, rhs: &Self) -> Self

Compute the interval of possible results if one multiplies a value in self with a value in rhs.

pub fn shift_left(&self, rhs: &Self) -> Self

Compute the resulting interval after a left shift operation. The result is only exact if the rhs interval contains exactly one value.

impl IntervalDomain

pub fn new(start: Bitvector, end: Bitvector) -> Self

Create a new interval domain with the given bounds.

Both start and end are inclusive, i.e. contained in the interval. The widening hints are set to None and the stride is set to 1 if start != end.

pub fn equal_as_value_sets(&self, other: &IntervalDomain) -> bool

Returns true if the two intervals represent the same value sets. This function ignores differences in the widening hints of the two intervals.

pub fn update_widening_lower_bound(&mut self, bound: &Option<Bitvector>)

If bound is more exact/restrictive than the current lower bound of self, set the lower bound to bound. Otherwise keep the old lower bound.

pub fn update_widening_upper_bound(&mut self, bound: &Option<Bitvector>)

If bound is more exact/restrictive than the current upper bound of self, set the upper bound to bound. Otherwise keep the old upper bound.

pub fn signed_merge(&self, other: &IntervalDomain) -> IntervalDomain

Merge as signed intervals without performing widenings.

pub fn signed_merge_and_widen(&self, other: &IntervalDomain) -> IntervalDomain

Merge as signed intervals and perform widening if necessary.

Widening Strategy

The widening delay

Each interval has a widening_delay counter, which denotes the length of the interval after the last time that widening was performed. For operations with more than one input, the widening delay is set to the maximum of the input widening delays. The only exception to this is the IntervalDomain::intersect() method, which may lower the value of the widening delay.

When to widen

If the merged interval equals one of the input intervals as value sets, do not perform widening. Else widening is performed if and only if the length of the interval is greater than the widening delay plus the stride of the interval.

How to widen

If no suitable widening bounds for widening exist, widen to the Top value. If exactly one widening bound exists, widen up to the bound, but do not perform widening in the other direction of the interval. If widening bounds for both directions exist, widen up to the bounds in both directions.

After that the widening_delay is set to the length of the resulting interval.

pub fn zero_extend(self, width: ByteSize) -> IntervalDomain

Zero-extend the values in the interval to the given width.

pub fn sign_extend(self, width: ByteSize) -> Self

Sign-extend the values in the interval to the given width.

pub fn fits_into_size(&self, size: ByteSize) -> bool

Check whether all values in the interval are representable by bitvectors of the given size. Does not check whether this is also true for the widening hints.