Struct cwe_checker_lib::abstract_domain::MemRegionData

pub struct MemRegionData<T: AbstractDomain + SizedDomain + HasTop + Debug> { /* fields omitted */ }

The internal data of a memory region. See the description of MemRegion for more.

Implementations

impl<T: AbstractDomain + SizedDomain + HasTop + Debug> MemRegionData<T>

pub fn new(address_bytesize: ByteSize) -> MemRegionData<T>

create a new, empty MemRegion

pub fn get_address_bytesize(&self) -> ByteSize

Get the bitsize of pointers for the address space that the memory region belongs to.

pub fn clear_offset_interval(
&mut self,
start: i64,
end: i64,
value_size: ByteSize
)

Clear all values that might be overwritten if one writes a value with byte size value_size to an offset contained in the interval from start to end (both bounds included in the interval).

This represents the effect of writing an arbitrary value (with known byte size) to an arbitrary offset contained in the interval.

pub fn add(&mut self, value: T, position: Bitvector)

Add a value to the memory region.

pub fn insert_at_byte_index(&mut self, value: T, position: i64)

Insert a value into the memory region at the given position. The position is the index (in bytes) in the memory region.

pub fn get(&self, position: Bitvector, size_in_bytes: ByteSize) -> T

Get the value at the given position. If there is no value at the position or the size of the element is not the same as the provided size, return T::new_top().

pub fn get_unsized(&self, position: Bitvector) -> Option<T>

Get the value at the given position regardless of the value size. Return None if there is no value at that position in the memory region.

pub fn remove(&mut self, position: Bitvector, size_in_bytes: Bitvector)

Remove all elements intersecting the provided interval.

pub fn merge(&self, other: &MemRegionData<T>) -> MemRegionData<T>

Merge two memory regions.

Values at the same position and with the same size get merged via their merge function. Other values are not added to the merged region, because they could be anything in at least one of the two regions.

pub fn iter(&self) -> Iter<'_, i64, T>

Get an iterator over all elements together with their offset into the memory region.

pub fn values(&self) -> Values<'_, i64, T>

Get an iterator over all values in the memory region

pub fn entry_map(&self) -> &BTreeMap<i64, T>

Get the map of all elements including their offset into the memory region.

pub fn values_mut(&mut self) -> ValuesMut<'_, i64, T>

Get an iterator over all values in the memory region for in-place manipulation. Note that one can changes values to Top using the iterator. These values should be removed from the memory region using clear_top_values().

pub fn clear_top_values(&mut self)

Remove all values representing the Top element from the internal value store, as these should not be saved in the internal representation.