Trait cwe_checker_lib::abstract_domain::DomainInsertion

pub trait DomainInsertion {
    fn append_string_domain(&self, string_domain: &Self) -> Self;

    fn create_integer_domain() -> Self;

    fn create_char_domain() -> Self;

    fn create_float_value_domain() -> Self;

    fn create_pointer_value_domain() -> Self;

    fn create_top_value_domain() -> Self;

    fn create_empty_string_domain() -> Self;
}

A set of functions that all abstract string domains should implement.

Required methods

fn append_string_domain(&self, string_domain: &Self) -> Self

Inserts a string domain at a certain position if order is considered.

fn create_integer_domain() -> Self

Creates a string domain with characters that usually appear in an integer value.

fn create_char_domain() -> Self

Creates a string domain with characters that usually appear in a char value.

fn create_float_value_domain() -> Self

Creates a string domain with characters that usually appear in a float value.

fn create_pointer_value_domain() -> Self

Creates a string domain with characters that usually appear in a String value.

fn create_top_value_domain() -> Self

Creates a top value of the currently used domain.

fn create_empty_string_domain() -> Self

Creates an empty string domain.