Trait cwe_checker_lib::abstract_domain::HasTop

pub trait HasTop {
    // Required method
    fn top(&self) -> Self;
}

An abstract domain implementing this trait has a global maximum, i.e. a Top element.

Required Methods

fn top(&self) -> Self

Return an instance of the Top element.

Since an abstract domain type may represent a whole family of abstract domains, this function takes an instance of the domain as a parameter, so it can return the Top element of the same family member that the provided instance belongs to.

Object Safety

This trait is not object safe.