Trait cwe_checker_lib::abstract_domain::SizedDomain

pub trait SizedDomain {
    fn bytesize(&self) -> ByteSize;

    fn new_top(bytesize: ByteSize) -> Self;
}

A trait for types representing values with a fixed size (in bytes).

For abstract domains, the bytesize is a parameter of the domain itself, i.e. you cannot merge values of different bytesizes, since they lie in different posets (one for each bytesize).

Required methods

fn bytesize(&self) -> ByteSize

Return the size of the represented value in bytes.

fn new_top(bytesize: ByteSize) -> Self

Return a new top element with the given bytesize. The function is expected to panic if the type in question does not also implement the HasTop trait.