Trait cwe_checker_lib::abstract_domain::TryToBitvec

pub trait TryToBitvec {
    fn try_to_bitvec(&self) -> Result<Bitvector, Error>;

    fn try_to_offset(&self) -> Result<i64, Error> { ... }
}

A conversion trait for abstract domains that can represent register values.

Required methods

fn try_to_bitvec(&self) -> Result<Bitvector, Error>

If self represents a single absolute value, return it. In all other cases return an error.

Loading content...
Provided methods

fn try_to_offset(&self) -> Result<i64, Error>

If self represents a single absolute value, try to convert it to a signed integer and return it. Else return an error. Note that the conversion loses information about the bytesize of the value.

Loading content...