Module cwe_checker_lib::analysis

Modules necessary for graph-based and fixpoint-based analyses, as well as analyses depending on these modules.

Modules

backward_interprocedural_fixpoint
Creating and computing backward interprocedural fixpoint problems.

dead_variable_elimination
This module contains a fixpoint computation to compute alive (resp. dead) variables and a function to remove dead assignments from a project.

fixpoint
Creating and computing generic fixpoint computations.

forward_interprocedural_fixpoint
Creating and computing forward interprocedural fixpoint problems.

graph
Generate control flow graphs out of a program term.

interprocedural_fixpoint_generic
Types and functions shared between the implementations of forward and backward interprocedural fixpoint computations.

pointer_inference
A fixpoint algorithm analyzing all memory accesses in a program.