Module cwe_checker_lib::analysis::forward_interprocedural_fixpoint

Creating and computing forward interprocedural fixpoint problems.

General notes

This module supports computation of fixpoint problems on the control flow graphs generated by the graph module.

To compute a generalized fixpoint problem, first construct a context object implementing the Contexttrait. Use it to construct a Computation object. The Computation object provides the necessary methods for the actual fixpoint computation.

Structs

GeneralizedContext
This struct is a wrapper to create a general fixpoint context out of an interprocedural fixpoint context.

Traits

Context
The context for an interprocedural fixpoint computation.

Functions

create_computation
Generate a new computation from the corresponding context and an optional default value for nodes.