Function cwe_checker_lib::analysis::function_signature::compute_function_signatures

pub fn compute_function_signatures<'a>(
    project: &'a Project,
    graph: &'a Graph<'_>
) -> (BTreeMap<Tid, FunctionSignature>, Vec<LogMessage>)

Compute the function signatures for all functions in the project.

Returns a map from the function TIDs to their signatures, and a list of log and debug messages recorded during the computation of the signatures.

For more information on the used algorithm see the module-level documentation.