Struct cwe_checker_lib::analysis::function_signature::AccessPattern

pub struct AccessPattern { /* private fields */ }

Access flags to track different kind of access/usage patterns of a variable.

Implementations

impl AccessPattern

pub fn new() -> Self

Generate a new AccessPattern object with none of the access flags set.

pub fn new_unknown_access() -> Self

Generate a new AccessPattern object with all access flags set to true (to model unknown access).

pub fn with_read_flag(self) -> Self

Set the access flag for read access and return self.

pub fn with_dereference_flag(self) -> Self

Set the access flag for immutable pointer dereference and return self.

pub fn with_mutably_dereferenced_flag(self) -> Self

Set the access flag for pointer dereference with write access to the pointer target and return self.

pub fn set_dereference_flag(&mut self)

Set the access flag for immutable pointer dereference.

pub fn set_read_flag(&mut self)

Set the access flag for read access.

pub fn set_mutably_dereferenced_flag(&mut self)

Set the access flag for pointer dereference (with write access to the target of the pointer).

pub fn set_unknown_access_flags(&mut self)

Set all access flags to indicate that any kind of access to the variable may have occured.

pub fn is_accessed(&self) -> bool

Returns true if any of the access flags is set.

pub fn is_dereferenced(&self) -> bool

Returns true if the dereferenced or mutably dereferenced access flag is set.

pub fn is_mutably_dereferenced(&self) -> bool

Returns true if the mutably dereferenced access flag is set.