Enum cwe_checker_lib::analysis::graph::Node

pub enum Node<'a> {
    BlkStart(&'a Term<Blk>, &'a Term<Sub>),
    BlkEnd(&'a Term<Blk>, &'a Term<Sub>),
    CallReturn {
        call: (&'a Term<Blk>, &'a Term<Sub>),
        return_: (&'a Term<Blk>, &'a Term<Sub>),
    },
    CallSource {
        source: (&'a Term<Blk>, &'a Term<Sub>),
        target: (&'a Term<Blk>, &'a Term<Sub>),
    },
}

The node type of an interprocedural control flow graph

Each node carries a pointer to its associated block with it. For CallReturnnodes the associated blocks are both the CallSourceblock (containing the call instruction) and the returning-from block (containing the return instruction).

For CallSourcenodes the associated block is the callsite block (source) and the target block of the call.

Basic blocks are allowed to be contained in more than one Sub. In the control flow graph such basic blocks occur once per subroutine they are contained in. For this reason, the nodes also carry a pointer to the corresponding subroutine with them to allow unambigous node identification.

Variants

BlkStart(&'a Term<Blk>, &'a Term<Sub>)

A node corresponding to the start of a basic block, i.e. to the point in time just before the execution of the block.

BlkEnd(&'a Term<Blk>, &'a Term<Sub>)

A node corresponding to the end of the basic block, i.e. to the point in time just after the execution of all Def instructions in the block but before execution of the jump instructions at the end of the block.

CallReturn

Fields

call: (&'a Term<Blk>, &'a Term<Sub>)

The block containing the callsite of the call.

return_: (&'a Term<Blk>, &'a Term<Sub>)

The block that the called functions returns to.

An artificial node. See the module-level documentation for more information.

CallSource

Fields

source: (&'a Term<Blk>, &'a Term<Sub>)

The block containing the callsite of the call

target: (&'a Term<Blk>, &'a Term<Sub>)

The block containing the target of the call, i.e. the first block of the target function.

An artificial node. See the module-level documentation for more information.

Implementations

impl<'a> Node<'a>

pub fn get_block(&self) -> &'a Term<Blk>

Get the block corresponding to the node for BlkStart and BlkEnd nodes. panics if called on a CallReturn node.

pub fn get_sub(&self) -> &'a Term<Sub>

Get the sub corresponding to the node for BlkStart and BlkEnd nodes. panics if called on a CallReturn node.