Module cwe_checker_lib::analysis::pointer_inference

A fixpoint algorithm analyzing all memory accesses in a program.

The goal of the pointer inference analysis is to keep track of all memory objects and pointers that the program knows about at specific program points during execution. If the Memory check is enabled, then the analysis also detects and reports possible memory management errors, like “Use after free”-errors, to the user. The result of the pointer inference analysis is also used as input for other analyses, which allows them to keep track of memory objects and pointers.

The Memory Check

If the Memory check is enabled, the pointer inference detects instances of the following CWEs:

CWE-119 Buffer Overflow (generic case)

CWE-125 Buffer Overflow: Out-of-bounds Read

CWE-415: Double Free

CWE-416: Use After Free

CWE-787: Buffer Overflow: Out-of-bounds Write

The analysis operates on a best-effort basis. In cases where we cannot know whether an error is due to an error in the memory management of the program under analysis or due to inexactness of the pointer inference analysis itself, we try to treat is as the more likely (but not necessarily true) case of the two.

See the Config struct for configurable analysis parameters.

Modules

object
This module contains the definition of the abstract memory object type.

Structs

Config
Configurable parameters for the analysis.

PointerInference
A wrapper struct for the pointer inference computation object.

State
Contains all information known about the state of a program at a specific point of time.

Statics

CWE_MODULE
The name and version number of the “Memory” CWE check.

Functions

extract_pi_analysis_results
The entry point for the memory analysis check. Does not actually compute anything but just extracts the results of the already computed pointer inference analysis.

run
Compute the pointer inference analysis and return its results.

Type Definitions

Data
The abstract domain type for representing register values.

ValueDomain
The abstract domain to use for absolute values.