Function cwe_checker_lib::analysis::pointer_inference::run

pub fn run<'a>(
    project: &'a Project, 
    runtime_memory_image: &'a RuntimeMemoryImage, 
    control_flow_graph: &'a Graph<'a>, 
    config: Config, 
    print_debug: bool
) -> PointerInference<'a>

Compute the pointer inference analysis and return its results.

If print_debug is set to true print debug information to stdout. Note that the format of the debug information is currently unstable and subject to change.