Struct cwe_checker_lib::analysis::pointer_inference::object::AbstractObjectInfo

pub struct AbstractObjectInfo {
    pub is_unique: bool,
    // some fields omitted
}

The abstract object info contains all information that we track for an abstract object.

Fields
is_unique: bool

Tracks whether this may represent more than one actual memory object.

Implementations

impl AbstractObjectInfo

pub fn new(type_: ObjectType, address_bytesize: ByteSize) -> AbstractObjectInfo

Create a new abstract object with known object type and address bitsize

pub fn set_lower_index_bound(&mut self, lower_bound: BitvectorDomain)

Set the lower index bound that is still considered to be contained in the abstract object.

pub fn set_upper_index_bound(&mut self, upper_bound: BitvectorDomain)

Set the upper index bound that is still considered to be contained in the abstract object.

pub fn access_contained_in_bounds(
&self,
offset: &ValueDomain,
size: ByteSize
) -> bool

Check whether a memory access to the abstract object at the given offset and with the given size of the accessed value is contained in the bounds of the memory object. If offset contains more than one possible index value, then only return true if the access is contained in the abstract object for all possible offset values.

pub fn get_value(&self, offset: Bitvector, bytesize: ByteSize) -> Data

Read the value at the given offset of the given size (in bits, not bytes) inside the memory region.

pub fn set_value(
&mut self,
value: Data,
offset: &ValueDomain
) -> Result<(), Error>

Write a value at the given offset to the memory region.

If the abstract object is not unique (i.e. may represent more than one actual object), merge the old value at the given offset with the new value.

pub fn merge_value(&mut self, value: Data, offset: &ValueDomain)

Merge value at position offset with the value currently saved at that position.

pub fn get_referenced_ids_overapproximation(
&self
) -> &BTreeSet<AbstractIdentifier>

Get all abstract IDs that the object may contain pointers to. This yields an overapproximation of possible pointer targets.

pub fn get_referenced_ids_underapproximation(
&self
) -> BTreeSet<AbstractIdentifier>

Get all abstract IDs for which the object contains pointers to. This yields an underapproximation of pointer targets, since the object may contain pointers that could not be tracked by the analysis.

pub fn replace_abstract_id(
&mut self,
old_id: &AbstractIdentifier,
new_id: &AbstractIdentifier,
offset_adjustment: &ValueDomain
)

For pointer values replace an abstract identifier with another one and add the offset_adjustment to the pointer offsets. This is needed to adjust stack pointers on call and return instructions.

pub fn set_state(&mut self, new_state: ObjectState)

If self.is_unique==true, set the state of the object. Else merge the new state with the old.

pub fn remove_ids(&mut self, ids_to_remove: &BTreeSet<AbstractIdentifier>)

Remove the provided IDs from the target lists of all pointers in the memory object. Also remove them from the pointer_targets list.

pub fn get_state(&self) -> ObjectState

Get the state of the memory object.

pub fn get_object_type(&self) -> Option<ObjectType>

Get the type of the memory object.

pub fn assume_arbitrary_writes(
&mut self,
additional_targets: &BTreeSet<AbstractIdentifier>
)

Invalidates all memory and adds the additional_targets to the pointer targets. Represents the effect of unknown write instructions to the object which may include writing pointers to targets from the additional_targets set to the object.

pub fn mark_as_freed(&mut self) -> Result<(), Error>

Mark the memory object as freed. Returns an error if a possible double free is detected or the memory object may not be a heap object.

pub fn mark_as_maybe_freed(&mut self) -> Result<(), Error>

Mark the memory object as possibly (but not definitely) freed. Returns an error if the object was definitely freed before or if the object may not be a heap object.

impl AbstractObjectInfo

pub fn to_json_compact(&self) -> Value

Get a more compact json-representation of the abstract object. Intended for pretty printing, not useable for serialization/deserialization.