Struct cwe_checker_lib::analysis::pointer_inference::State

pub struct State {
    pub memory: AbstractObjectList,
    pub stack_id: AbstractIdentifier,
    /* private fields */
}

Contains all information known about the state of a program at a specific point of time.

Fields
memory: AbstractObjectList

The list of all known memory objects.

stack_id: AbstractIdentifier

The abstract identifier of the current stack frame. It points to the base of the stack frame, i.e. only negative offsets point into the current stack frame.

Implementations

impl State

pub fn get_register(&self, variable: &Variable) -> Data

Get the value of a register or Top() if no value is known.

pub fn set_register(&mut self, variable: &Variable, value: Data)

Set the value of a register.

pub fn handle_register_assign(
&mut self,
target: &Variable,
expression: &Expression
)

Evaluate expression on the given state and write the result to the target register.

pub fn store_value(
&mut self,
address: &Data,
value: &Data,
global_memory: &RuntimeMemoryImage
) -> Result<(), Error>

Store value at the given address.

pub fn write_to_address(
&mut self,
address: &Expression,
value: &Data,
global_memory: &RuntimeMemoryImage
) -> Result<(), Error>

Write a value to the address one gets when evaluating the address expression.

pub fn handle_store(
&mut self,
address: &Expression,
value: &Expression,
global_memory: &RuntimeMemoryImage
) -> Result<(), Error>

Evaluate the store instruction, given by its address and value expressions, and modify the state accordingly.

pub fn load_value(
&self,
address: &Expression,
size: ByteSize,
global_memory: &RuntimeMemoryImage
) -> Result<Data, Error>

Evaluate the given load instruction and return the data read on success.

pub fn handle_load(
&mut self,
var: &Variable,
address: &Expression,
global_memory: &RuntimeMemoryImage
) -> Result<(), Error>

Handle a load instruction by assigning the value loaded from the address given by the address expression to var.

pub fn eval(&self, expression: &Expression) -> Data

Evaluate the value of an expression in the current state

pub fn eval_parameter_arg(
&self,
parameter: &Arg,
global_memory: &RuntimeMemoryImage
) -> Result<Data, Error>

Evaluate the value of a parameter of an extern symbol for the given state.

pub fn check_def_for_null_dereferences(
&mut self,
def: &Term<Def>
) -> Result<bool, Error>

Check whether the given def could result in a memory access through a NULL pointer.

If no NULL pointer dereference is detected then Ok(false) is returned. If a NULL pointer dereference is detected, try to specialize the state so that address_expr cannot result in a NULL pointer anymore. If that succeeds, Ok(true) is returned. If that would result in an unsatisfiable state, an error is returned.

impl State

pub fn add_directly_reachable_ids_to_id_set(
&self,
ids: BTreeSet<AbstractIdentifier>
) -> BTreeSet<AbstractIdentifier>

Search (recursively) through all memory objects referenced by the given IDs and add all IDs reachable through concrete pointers contained in them to the set of IDs.

This uses an underapproximation of the referenced IDs of a memory object, i.e. IDs may be missing if the analysis lost track of the corresponding pointer.

pub fn add_recursively_referenced_ids_to_id_set(
&self,
ids: BTreeSet<AbstractIdentifier>
) -> BTreeSet<AbstractIdentifier>

Search (recursively) through all memory objects referenced by the given IDs and add all IDs contained in them to the set of IDs.

This uses an overapproximation of the referenced IDs of a memory object, i.e. for a memory object it may add IDs as possible references where the corresponding reference is not longer present in the memory object.

pub fn add_param_object_from_callee(
&mut self,
param_object: AbstractObject,
param_value_at_call: &Data
) -> Result<(), Error>

Add the given param_object from the callee state to self (where self represents the state after returning from the callee).

param_value_at_call is the value that the parameter had at the callsite. It is assumed that all IDs contained in the param_object are already replaced with values relative to the caller.

If the param_object corresponds to a unique object in self then the contents of that object are overwritten with those of param_object. Else the contents are only merged with all possible caller objects, since the exact object that corresponds to the callee object is unknown.

impl State

pub fn specialize_by_expression_result(
&mut self,
expression: &Expression,
result: Data
) -> Result<(), Error>

Try to restrict the input variables of expression on self so that expression only evaluates to values represented by the given result.

If expression cannot evaluate to any value represented by self, return an error.

This function may restrict to upper bounds of possible values if the restriction cannot be made exact, i.e. after calling this function the state may still contain values for which expression does not evaluate to values represented by result.

impl State

pub fn new(stack_register: &Variable, function_tid: Tid) -> State

Create a new state that contains only one memory object corresponding to the stack. The stack offset will be set to zero.

pub fn from_fn_sig(
fn_sig: &FunctionSignature,
stack_register: &Variable,
function_tid: Tid
) -> State

Create a new state from a function signature.

The created state contains one memory object for the stack frame of the function and one memory object for each parameter that is dereferenced by the function (according to the function signature).

pub fn set_mips_link_register(
&mut self,
callee_tid: &Tid,
generic_pointer_size: ByteSize
) -> Result<(), Error>

Set the MIPS link register t9 to the address of the callee TID.

According to the System V ABI for MIPS the caller has to save the callee address in register t9 on a function call to position-independent code. This function manually sets t9 to the correct value to mitigate cases where t9 could not be correctly computed due to previous analysis errors.

Returns an error if the callee address could not be parsed (e.g. for UNKNOWN addresses).

pub fn clear_non_callee_saved_register(
&mut self,
callee_saved_register: &[Variable]
)

Clear all non-callee-saved registers from the state. This automatically also removes all virtual registers. The parameter is a list of callee-saved register names.

pub fn clear_stack_parameter(
&mut self,
extern_call: &ExternSymbol,
global_memory: &RuntimeMemoryImage
) -> Result<(), Error>

Mark those parameter values of an extern function call, that are passed on the stack, as unknown data (since the function may modify them).

pub fn remove_unreferenced_objects(&mut self)

Remove all objects that cannot longer be reached by any known pointer. This does not remove objects, where some caller may still know a pointer to the object.

The function uses an underapproximation of all possible pointer targets contained in a memory object. This keeps the number of tracked objects reasonably small.

pub fn remove_non_callee_saved_register(&mut self, cconv: &CallingConvention)

Remove all knowledge about the contents of non-callee-saved registers from the state.

pub fn get_fn_tid(&self) -> &Tid

Get the Tid of the function that this state belongs to.

impl State

pub fn to_json_compact(&self) -> Value

Get a more compact json-representation of the state. Intended for pretty printing, not useable for serialization/deserialization.