Struct cwe_checker_lib::analysis::string_abstraction::StringAbstraction

pub struct StringAbstraction<'a, T: AbstractDomain + DomainInsertion + HasTop + Eq + From<String>> { /* private fields */ }

A wrapper struct for the string abstraction computation object.

Implementations

impl<'a, T: AbstractDomain + DomainInsertion + HasTop + Eq + From<String>> StringAbstraction<'a, T>

pub fn new( project: &'a Project, control_flow_graph: &'a Graph<'a>, pointer_inference_results: &'a PointerInferenceComputation<'a>, config: Config ) -> StringAbstraction<'a, T>

Generate a new string abstraction computation for a project.

pub fn compute(&mut self)

Compute the fixpoint of the string abstraction analysis. Has a max_steps bound for the fixpoint algorithm to prevent infinite loops.

pub fn get_computation( &self ) -> &Computation<GeneralizedContext<'a, Context<'a, T>>>

Get the string abstraction computation.

pub fn get_graph(&self) -> &Graph<'_>

Get the underlying graph of the computation.

pub fn get_context(&self) -> &Context<'a, T>

Get the context object of the computation.

pub fn get_node_value(&self, node_id: NodeIndex) -> Option<&NodeValue<State<T>>>

Get the value associated to a node in the computed fixpoint (or intermediate state of the algorithm if the fixpoint has not been reached yet). Returns None if no value is associated to the Node.