Module cwe_checker_lib::analysis::vsa_results

This module provides the VsaResult trait which defines an interface for the results of analyses similar to a value set analysis.

Traits

HasVsaResult

Trait for types that provide access to the result of a value set analysis.

VsaResult

A trait providing an interface for accessing the results of a value set analysis. Note that the returned values may be any type of information associated with values at certain program points, i.e. the trait can also be used for other analyses than just value set analyses.