Trait cwe_checker_lib::analysis::vsa_results::HasVsaResult

pub trait HasVsaResult<T> {
    // Required method
    fn vsa_result(&self) -> &impl VsaResult<ValueDomain = T>;
}

Trait for types that provide access to the result of a value set analysis.

The generic type parameter can be used to implement this trait multiple times, i.e., the same type can provide access to VSA results with different value domains.

Required Methods

fn vsa_result(&self) -> &impl VsaResult<ValueDomain = T>

Converts a reference to Self into a reference to a type that implements VsaResult with ValueDomain T.

Object Safety

This trait is not object safe.