Module cwe_checker_lib::checkers::cwe_337

This module implements a check for CWE-337: Predictable Seed in Pseudo-Random Number Generator (PRNG)

The use of predictable seeds significantly reduces the number of possible seeds that an attacker would need to test in order to predict which random numbers will be generated by the PRNG.

See https://cwe.mitre.org/data/definitions/337.html for a detailed description.

How the check works

Using dataflow analysis we search for an execution path where the result of a time source, like time, is used as an argument to a PRNG seeding function, like srand.

Symbols configurable in config.json

Both the sources of predictable seeds and the seeding functions can be configured using the sources and seeding_functions respectively.

Structs

Config

The configuration struct

Context

The Context struct for the forward_interprocedural_fixpoint algorithm.

Statics

CWE_MODULE

The module name and version

Functions

check_cwe

This check checks if a return value of any of the sources (as determined by the config file) is used as a direct parameter of any of the sinks (as determined by the config file). Currently, this is only used to detect whether a call of time leads into a call of srand, only tracks registers and not memory locations and stops at any call.