Enum cwe_checker_lib::checkers::cwe_476::Taint

pub enum Taint {
    Tainted(ByteSize),
    Top(ByteSize),
}

An abstract domain representing a value that is either tainted or not.

Note that the merge-function does not respect the partial order that is implied by the naming scheme of the variants! In fact the whole analysis does not enforce any partial order for this domain. This means that in theory the fixpoint computation may not actually converge to a fixpoint, but in practice the analysis can make more precise decisions whether a value should be tainted or not.

Variants

Tainted(ByteSize)

A tainted value of a particular bytesize.

Top(ByteSize)

An untainted value of a particular bytesize

Implementations

impl Taint

pub fn is_tainted(&self) -> bool

Checks whether the given value is in fact tainted.