Module cwe_checker_lib::checkers::cwe_78

This module implements a check for CWE-78: Improper Neutralization of Special Elements used in an OS Command (‘OS Command Injection’).

The software constructs all or part of an OS command using externally-influenced input from an upstream component, but it does not neutralize or incorrectly neutralizes special elements that could modify the intended OS command when it is sent to a downstream component.

See https://cwe.mitre.org/data/definitions/78.html for a detailed description.

How the check works

Using backward dataflow analysis we search for an executation path from a system call parameter (string) to an user input to identify possible command injections.

To find relevant string related functions, such as sprintf, it is assumed that the first input parameter points to the memory position that will be used as the return location. (e.g. char *strcat(char *dest, const char *src) where ‘char *dest’ will contain the return value)

For instance:

MOV RAX, qword ptr [RBP + local_10]
MOV RDI, RAX                        // RDI is the first input parameter for the strcat call and it points to [RBP + local_10]
CALL strcat
MOV RAX, qword ptr [RBP + local_10] // In the backwards analysis [RBP + local_10] will be tainted and it contains the return value

Symbols configurable in config.json

The symbols are the functions which

make system calls (e.g. system)

manipulate strings (e.g. sprintf, strcat, memcpy, etc.)

take user input (e.g. scanf)

False Positives

The input comes from the user but proper sanitization was not detected by the analysis even though it exists.

The input comes from the user but the format string’s input format could not be distinguished as non-string input.

False Negatives

Missing Taints due to lost track of pointer targets

Non tracked function parameters cause incomplete taints that could miss possible dangerous inputs

Structs

BlockMaps

block_first_def_set: - A set containing a given Def as the first Def of the block. The keys are of the form (Def-TID, Current-Sub-TID) to distinguish the nodes for blocks contained in more than one function.

block_start_last_def_map: - A map to get the node index of the BlkStart node containing a given Def as the last Def of the block. The keys are of the form (Def-TID, Current-Sub-TID) to distinguish the nodes for blocks contained in more than one function.

jmp_to_blk_end_node_map: - A map to get the node index of the BlkEnd node containing a given Jmp. The keys are of the form (Jmp-TID, Current-Sub-TID) to distinguish the nodes for blocks contained in more than one function.

Config
The configuration struct

SymbolMaps

string_symbols:

Maps the TID of an extern string related symbol to the corresponding extern symbol struct.

user_input_symbols:

Maps the TID of an extern symbol that take input from the user to the corresponding extern symbol struct.

extern_symbol_map:

Maps the TID of an extern symbol to the extern symbol struct.

format_string_index:

Maps a symbol name to the index of its format string parameter.

Statics

CWE_MODULE
The module name and version

Functions

check_cwe
This check searches for system calls and sets their parameters as taint source if available. Then the fixpoint computation is executed and its result may generate cwe warnings if the parameters can be tracked back to user inputs