Struct cwe_checker_lib::intermediate_representation::Term

pub struct Term<T> {
    pub tid: Tid,
    pub term: T,
}

A term is an object inside a binary with an address and an unique ID (both contained in the tid).

Fields
tid: Tid

The term identifier, which also contains the address of the term

term: T

The object

Implementations

impl Term<Def>

pub fn check_for_zero_extension(
&self,
output_name: String,
output_sub_register: String
) -> Option<Tid>

This function checks whether the instruction is a zero extension of the overwritten sub register of the previous instruction. If so, returns its TID

pub fn substitute_input_var(
&mut self,
input_var: &Variable,
replace_with_expression: &Expression
)

Substitute every occurence of input_var in the address and value expressions with replace_with_expression. Does not change the target variable of assignment- and load-instructions.

impl Term<Blk>

pub fn remove_nonexisting_indirect_jump_targets(
&mut self,
known_block_tids: &HashSet<Tid>
) -> Result<(), Vec<LogMessage>>

Remove indirect jump target addresses for which no corresponding target block exists. Return an error message for each removed address.

pub fn propagate_input_expressions(&mut self)

Wherever possible, substitute input variables of expressions with the input expression that defines the input variable.

Note that substitution is only possible if the input variables of the input expression itself did not change since the definition of said variable.

The expression propagation allows the corresponding Project normalization pass to further simplify the generated expressions and allows more dead stores to be removed during dead variable elimination.

pub fn merge_def_assignments_to_same_var(&mut self)

Merge subsequent assignments to the same variable to a single assignment to that variable.

The value expressions of merged assignments can often be simplified later on in the corresponding Project normalization pass.

impl Term<Sub>

pub fn into_ir_sub_term(self, generic_pointer_size: ByteSize) -> Term<IrSub>

Convert a Sub term in the P-Code representation to a Sub term in the intermediate representation. The conversion also repairs the order of the basic blocks in the blocks array of the Sub in the sense that the first block of the array is required to also be the function entry point after the conversion.