Function cwe_checker_lib::utils::arguments::parse_format_string_parameters

pub fn parse_format_string_parameters(
    format_string: &str, 
    datatype_properties: &DatatypeProperties
) -> Result<Vec<(Datatype, ByteSize)>, Error>

Parses the format string parameters using a regex, determines their data types, and calculates their positions (register or memory).