Function cwe_checker_lib::utils::symbol_utils::get_calls_to_symbols

pub fn get_calls_to_symbols<'a, 'b>(
    sub: &'a Term<Sub>, 
    symbols: &'b HashMap<&'a Tid, &'a str>
) -> Vec<(&'a str, &'a Tid, &'a str)>

Match direct calls’ target tids in the program’s subroutines with the tids of the external symbols given to the function. When a match was found, add a triple of (caller name, callsite tid, callee name) to a vector. Lastly, return the vector with all callsites of all given external symbols.