Struct cwe_checker_lib::abstract_domain::DomainMap

pub struct DomainMap<K, V, S>where
    K: PartialOrd + Ord + Clone,
    V: AbstractDomain,
    S: MapMergeStrategy<K, V>,{ /* private fields */ }

A DomainMap<Key, Value, MapMergeStrategy> is a wrapper type around a BTreeMap<Key, Value> where the Value` type is an abstract domain and the map itself is also an abstract domain.

For example, a map from registers to an abstract domain representing the contained values can be represented by a DomainMap.

A DomainMap has two main advantages over a regular BTreeMap:

The map itself is wrapped into an Arc<..> to enable cheap cloning of DomainMaps.

The DomainMap automatically implements the AbstractDomain trait according to the provided MapMergeStrategy used for merging two maps.

Since a DomainMap implements the Deref and DerefMut traits with target the inner BTreeMap, it can be used just like a BTreeMap.