Trait cwe_checker_lib::abstract_domain::AbstractDomain

pub trait AbstractDomain: Sized + Eq + Clone {
    // Required methods
    fn merge(&self, other: &Self) -> Self;
    fn is_top(&self) -> bool;
}

The main trait describing an abstract domain.

Each abstract domain is partially ordered. Abstract domains of the same type can be merged.

Required Methods

fn merge(&self, other: &Self) -> Self

Return an upper bound (with respect to the partial order on the domain) for the two inputs self and other.

fn is_top(&self) -> bool

Returns whether the element represents the top element (i.e. maximal with respect to the partial order) or not. If a domain has no maximal element, this function should always return false.