Struct cwe_checker_lib::analysis::fixpoint::Computation

pub struct Computation<T: Context> { /* private fields */ }

The computation struct contains an intermediate result of a fixpoint computation and provides methods for continuing the fixpoint computation or extracting the (intermediate or final) results.

Usage

let mut computation = Computation::new(context, optional_default_node_value);

// set starting node values with computation.set_node_value(..)
// ...

computation.compute();

// get the resulting node values
if let Some(node_value) = computation.get_node_value(node_index) {
    // ...
};

Implementations

impl<T: Context> Computation<T>

pub fn new(fp_context: T, default_value: Option<T::NodeValue>) -> Self

Create a new fixpoint computation from a fixpoint problem, the corresponding graph and a default value for all nodes if one should exists.

pub fn from_node_priority_list( fp_context: T, default_value: Option<T::NodeValue>, priority_sorted_nodes: Vec<NodeIndex> ) -> Self

Create a new fixpoint computation from a fixpoint problem, an optional default value and the list of nodes of the graph ordered by the priority for the worklist algorithm. The worklist algorithm will try to stabilize the nodes with a higher index in the priority_sorted_nodes array before those with a lower index.

pub fn get_node_value(&self, node: NodeIndex) -> Option<&T::NodeValue>

Get the value of a node.

pub fn set_node_value(&mut self, node: NodeIndex, value: T::NodeValue)

Set the value of a node and mark the node as not yet stabilized.

pub fn compute_with_max_steps(&mut self, max_steps: u64)

Compute the fixpoint of the fixpoint problem. Each node will be visited at most max_steps times. If a node does not stabilize after max_steps visits, the end result will not be a fixpoint but only an intermediate result of a fixpoint computation.

pub fn compute(&mut self)

Compute the fixpoint of the fixpoint problem. If the fixpoint algorithm does not converge to a fixpoint, this function will not terminate.

pub fn node_values(&self) -> &FnvHashMap<NodeIndex, T::NodeValue>

Get a reference to the internal map where one can look up the current values of all nodes

pub fn get_graph(&self) -> &DiGraph<T::NodeLabel, T::EdgeLabel>

Get a reference to the underlying graph

pub fn get_context(&self) -> &T

Get a reference to the underlying context object

pub fn has_stabilized(&self) -> bool

Returns True if the computation has stabilized, i.e. the internal worklist is empty.

pub fn get_worklist(&self) -> Vec<NodeIndex>

Return a list of all nodes which are marked as not-stabilized