Module cwe_checker_lib::analysis::function_signature

A fixpoint algorithm computing parameters of functions and their access patterns.

The fixpoint algorithm tracks the values of registers and the stack, although only stack accesses with known, constant offset are processed. Accesses to potential function parameters are collected together with the type of the access (is the value read, dereferenced for read access or dereferenced for write access).

Known limitations of the analysis:

The analysis is an overapproximation in the sense that it may generate more input parameters than actually exist in some cases.

Only registers that are potential parameter registers in the standard calling convention of the CPU architecture are considered as potential parameter registers. For functions that use other registers than those in the standard calling convention for parameter passing the results of this analysis will be wrong.

Parameters that are used as input values for variadic functions (e.g. sprintf) may be missed since detection of variadic function parameters is not yet implemented for this analysis.

If only a part (e.g. a single byte) of a stack parameter is accessed instead of the whole parameter then a duplicate stack parameter may be generated. A proper sanitation for this case is not yet implemented, although error messages are generated if such a case is detected.

For floating point parameter registers the base register is detected as a parameter, although only a smaller sub-register is the actual parameter in many cases. Also, if a function uses sub-registers of floating point registers as local variables, the registers may be incorrectly flagged as input parameters.

Structs

AccessPattern

Access flags to track different kind of access/usage patterns of a variable.

FunctionSignature

The signature of a function. Currently only contains information on the parameters of a function and their access patterns.

Functions

compute_function_signatures

Compute the function signatures for all functions in the project.