Module cwe_checker_lib::analysis::interprocedural_fixpoint_generic

Types and functions shared between the implementations of forward and backward interprocedural fixpoint computations.

Enums

NodeValue

NodeValue that can either be a single abstract value or a composition of the abstract value computed following an interprocedural call in the graph and of the abstract value before or after the call (depending on the direction of the fixpoint analysis). The CallFlowCombinator then allows for a merge of the values computed over both paths.

Functions

merge_option

Helper function to merge to values wrapped in Option<..>. Merges (Some(x), None) to Some(x).