Struct cwe_checker_lib::analysis::pointer_inference::PointerInference

pub struct PointerInference<'a> {
    pub collected_logs: (Vec<LogMessage>, Vec<CweWarning>),
    // some fields omitted
}

A wrapper struct for the pointer inference computation object.

Fields
collected_logs: (Vec<LogMessage>, Vec<CweWarning>)

The log messages and CWE warnings that have been generated during the pointer inference analysis.

Implementations

impl<'a> PointerInference<'a>

pub fn new(
project: &'a Project,
runtime_memory_image: &'a RuntimeMemoryImage,
control_flow_graph: &'a Graph<'a>,
config: Config,
log_sender: Sender<LogThreadMsg>
) -> PointerInference<'a>

Generate a new pointer inference compuation for a project.

pub fn compute(&mut self)

Compute the fixpoint of the pointer inference analysis. Has a max_steps bound for the fixpoint algorithm to prevent infinite loops.

pub fn print_yaml(&self)

Print results serialized as YAML to stdout

pub fn generate_compact_json(&self) -> Value

Generate a compacted json representation of the results. Note that this output cannot be used for serialization/deserialization, but is only intended for user output and debugging.

pub fn print_compact_json(&self)

Print a compacted json representation of the results to stdout. Note that this output cannot be used for serialization/deserialization, but is only intended for user output and debugging.

pub fn get_graph(&self) -> &Graph<'_>

Get the underlying graph of the computation.

pub fn get_context(&self) -> &Context<'_>

Get the context object of the computation.

pub fn get_node_value(&self, node_id: NodeIndex) -> Option<&NodeValue<State>>

Get the value associated to a node in the computed fixpoint (or intermediate state of the algorithm if the fixpoint has not been reached yet). Returns None if no value is associated to the Node.

pub fn compute_with_speculative_entry_points(&mut self, project: &Project)

Compute the results of the pointer inference fixpoint algorithm. Successively adds more functions as possible entry points to increase code coverage.