Struct cwe_checker_lib::analysis::pointer_inference::PointerInference

pub struct PointerInference<'a> {
    pub collected_logs: (Vec<LogMessage>, Vec<CweWarning>),
    /* private fields */
}

A wrapper struct for the pointer inference computation object. Also contains different analysis results computed through the fixpoint computation including generated log messages.

Fields
collected_logs: (Vec<LogMessage>, Vec<CweWarning>)

The log messages and CWE warnings that have been generated during the pointer inference analysis.

Implementations

impl<'a> PointerInference<'a>

pub fn new(
analysis_results: &'a AnalysisResults<'a>,
config: Config,
log_sender: Sender<LogThreadMsg>,
print_stats: bool
) -> PointerInference<'a>

Generate a new pointer inference computation for a project.

pub fn compute(&mut self, print_stats: bool)

Compute the fixpoint of the pointer inference analysis. Has a max_steps bound for the fixpoint algorithm to prevent infinite loops.

If print_stats is true then some extra log messages with statistics about the computation are generated.

pub fn print_yaml(&self)

Print results serialized as YAML to stdout

pub fn generate_compact_json(&self) -> Value

Generate a compacted json representation of the results. Note that this output cannot be used for serialization/deserialization, but is only intended for user output and debugging.

pub fn print_compact_json(&self)

Print a compacted json representation of the results to stdout. Note that this output cannot be used for serialization/deserialization, but is only intended for user output and debugging.

pub fn get_graph(&self) -> &Graph<'_>

Get the underlying graph of the computation.

pub fn get_context(&self) -> &Context<'_>

Get the context object of the computation.

pub fn get_node_value(&self, node_id: NodeIndex) -> Option<&NodeValue<State>>

Get the value associated to a node in the computed fixpoint (or intermediate state of the algorithm if the fixpoint has not been reached yet). Returns None if no value is associated to the Node.

pub fn get_state_at_jmp_tid(&self, jmp_tid: &Tid) -> Option<&State>

Get the state of the fixpoint computation at the block end node before the given jump instruction. This function only yields results after the fixpoint has been computed.

pub fn get_id_renaming_map_at_call_tid(
&self,
call_tid: &Tid
) -> Option<&BTreeMap<AbstractIdentifier, Data>>

Get the mapping from callee IDs to caller values for the given call. This function only yields results after the fixpoint has been computed.

Note that the maps may contain mappings from callee IDs to temporary caller IDs that get instantly removed from the caller since they are not referenced in any caller object.