Struct cwe_checker_lib::analysis::pointer_inference::State

pub struct State {
    pub memory: AbstractObjectList,
    pub stack_id: AbstractIdentifier,
    pub caller_stack_ids: BTreeSet<AbstractIdentifier>,
    pub ids_known_to_caller: BTreeSet<AbstractIdentifier>,
    // some fields omitted
}

Contains all information known about the state of a program at a specific point of time.

Fields
memory: AbstractObjectList

The list of all known memory objects.

stack_id: AbstractIdentifier

The abstract identifier of the current stack frame. It points to the base of the stack frame, i.e. only negative offsets point into the current stack frame.

caller_stack_ids: BTreeSet<AbstractIdentifier>

All known IDs of caller stack frames. Note that these IDs are named after the callsite, i.e. we can distinguish every callsite and for recursive functions the caller and current stack frames have different IDs.

Writes to the current stack frame with offset >= 0 are written to all caller stack frames. Reads to the current stack frame with offset >= 0 are handled as merge-read from all caller stack frames.

ids_known_to_caller: BTreeSet<AbstractIdentifier>

All IDs of objects that are known to some caller. This is an overapproximation of all object IDs that may have been passed as parameters to the function. The corresponding objects are not allowed to be deleted (even if no pointer to them exists anymore) so that after returning from a call the caller can recover their modified contents and the callee does not accidentally delete this information if it loses all pointers to an object.

Note that IDs that the callee should not have access to are not included here. For these IDs the caller can assume that the contents of the corresponding memory object were not accessed or modified by the call.

Implementations

impl State

pub fn get_register(&self, variable: &Variable) -> Data

Get the value of a register or Top() if no value is known.

pub fn get_register_by_name(&self, reg_name: &str) -> Option<Data>

Get the value of a register by its name.

Returns None if no value is set for the register.

pub fn set_register(&mut self, variable: &Variable, value: Data)

Set the value of a register.

pub fn handle_register_assign(
&mut self,
target: &Variable,
expression: &Expression
)

Evaluate expression on the given state and write the result to the target register.

pub fn store_value(
&mut self,
address: &Data,
value: &Data,
global_memory: &RuntimeMemoryImage
) -> Result<(), Error>

Store value at the given address.

pub fn write_to_address(
&mut self,
address: &Expression,
value: &Data,
global_memory: &RuntimeMemoryImage
) -> Result<(), Error>

Write a value to the address one gets when evaluating the address expression.

pub fn handle_store(
&mut self,
address: &Expression,
value: &Expression,
global_memory: &RuntimeMemoryImage
) -> Result<(), Error>

Evaluate the store instruction, given by its address and value expressions, and modify the state accordingly.

pub fn load_value(
&self,
address: &Expression,
size: ByteSize,
global_memory: &RuntimeMemoryImage
) -> Result<Data, Error>

Evaluate the given load instruction and return the data read on success.

pub fn handle_load(
&mut self,
var: &Variable,
address: &Expression,
global_memory: &RuntimeMemoryImage
) -> Result<(), Error>

Handle a load instruction by assigning the value loaded from the address given by the address expression to var.

pub fn eval(&self, expression: &Expression) -> Data

Evaluate the value of an expression in the current state

pub fn eval_parameter_arg(
&self,
parameter: &Arg,
stack_pointer: &Variable,
global_memory: &RuntimeMemoryImage
) -> Result<Data, Error>

Evaluate the value of a parameter of an extern symbol for the given state.

pub fn contains_access_of_dangling_memory(&mut self, def: &Def) -> bool

Check if an expression contains a use-after-free. If yes, mark the corresponding memory objects as flagged.

pub fn contains_out_of_bounds_mem_access(
&self,
def: &Def,
global_data: &RuntimeMemoryImage
) -> bool

Returns true if the given Def is a load or store instruction which may access a memory object outside its bounds.

pub fn pointer_contains_out_of_bounds_target(
&self,
data: &Data,
global_data: &RuntimeMemoryImage
) -> bool

Returns true if data is a pointer pointing outside of the bounds of a memory buffer. Does not check whether data may represent an out-of-bounds access to global memory, since this function assumes that all absolute values are not pointers.

pub fn is_stack_pointer_with_nonnegative_offset(&self, data: &Data) -> bool

Return true if data is a pointer to the current stack frame with a constant positive address, i.e. if it accesses a stack parameter (or the return-to address for x86) of the current function.

impl State

pub fn new(stack_register: &Variable, function_tid: Tid) -> State

Create a new state that contains only one memory object corresponding to the stack. The stack offset will be set to zero.

pub fn clear_non_callee_saved_register(
&mut self,
callee_saved_register_names: &[String]
)

Clear all non-callee-saved registers from the state. This automatically also removes all virtual registers. The parameter is a list of callee-saved register names.

pub fn clear_stack_parameter(
&mut self,
extern_call: &ExternSymbol,
stack_pointer_register: &Variable,
global_memory: &RuntimeMemoryImage
) -> Result<(), Error>

Mark those parameter values of an extern function call, that are passed on the stack, as unknown data (since the function may modify them).

pub fn replace_abstract_id(
&mut self,
old_id: &AbstractIdentifier,
new_id: &AbstractIdentifier,
offset_adjustment: &ValueDomain
)

Replace all occurences of old_id with new_id and adjust offsets accordingly. This is needed to replace stack/caller IDs on call and return instructions.

Example: Assume the old_id points to offset 0 in the corresponding memory object and the new_id points to offset -32. Then the offset_adjustment is -32. The offset_adjustment gets added to the base offset in self.memory.ids (so that it points to offset -32 in the memory object), while it gets subtracted from all pointer values (so that they still point to the same spot in the corresponding memory object).

pub fn remove_unreferenced_objects(&mut self)

Remove all objects that cannot longer be reached by any known pointer. This does not remove objects, where some caller may still know a pointer to the object.

The function uses an underapproximation of all possible pointer targets contained in a memory object. This keeps the number of tracked objects reasonably small.

pub fn add_directly_reachable_ids_to_id_set(
&self,
ids: BTreeSet<AbstractIdentifier>
) -> BTreeSet<AbstractIdentifier>

Search (recursively) through all memory objects referenced by the given IDs and add all IDs reachable through concrete pointers contained in them to the set of IDs.

This uses an underapproximation of the referenced IDs of a memory object, i.e. IDs may be missing if the analysis lost track of the corresponding pointer.

pub fn add_recursively_referenced_ids_to_id_set(
&self,
ids: BTreeSet<AbstractIdentifier>
) -> BTreeSet<AbstractIdentifier>

Search (recursively) through all memory objects referenced by the given IDs and add all IDs contained in them to the set of IDs.

This uses an overapproximation of the referenced IDs of a memory object, i.e. for a memory object it may add IDs as possible references where the corresponding reference is not longer present in the memory object.

pub fn merge_callee_stack_to_caller_stack(
&mut self,
callee_id: &AbstractIdentifier,
caller_id: &AbstractIdentifier,
offset_adjustment: &ValueDomain
)

Merge the callee stack with the caller stack.

This deletes the memory object corresponding to the callee_id and updates all other references pointing to the callee_id to point to the caller_id. The offset adjustment is handled as in replace_abstract_id.

Note that right now the content of the callee memory object is not merged into the caller memory object. In general this is the correct behaviour as the content below the stack pointer should be considered uninitialized memory after returning to the caller. However, an aggressively optimizing compiler or an unknown calling convention may deviate from this.

pub fn mark_mem_object_as_freed(
&mut self,
object_pointer: &PointerDomain<ValueDomain>
) -> Result<(), Vec<(AbstractIdentifier, Error)>>

Mark a memory object as already freed (i.e. pointers to it are dangling). If the object cannot be identified uniquely, all possible targets are marked as having an unknown status.

If this may cause double frees (i.e. the object in question may have been freed already), an error with the list of possibly already freed objects is returned.

pub fn remove_virtual_register(&mut self)

Remove all virtual register from the state. This should only be done in cases where it is known that no virtual registers can be alive.

Example: At the start of a basic block no virtual registers should be alive.

pub fn remove_other_caller_stack_ids(&mut self, caller_id: &AbstractIdentifier)

Recursively remove all caller_stack_ids not corresponding to the given caller.

pub fn readd_caller_objects(&mut self, caller_state: &State)

Add those objects from the caller_state to self, that are not known to self.

Since self does not know these objects, we assume that the current function could not have accessed them in any way during execution. This means they are unchanged from the moment of the call until the return from the call, thus we can simply copy their object-state from the moment of the call.

pub fn restore_callee_saved_register(
&mut self,
caller_state: &State,
cconv: &CallingConvention,
stack_register: &Variable
)

Restore the content of callee-saved registers from the caller state with the exception of the stack register.

This function does not check what the callee state currently contains in these registers. If the callee does not adhere to the given calling convention, this may introduce analysis errors! It will also mask cases where a callee-saved register was incorrectly modified (e.g. because of a bug in the callee).

pub fn remove_callee_saved_register(&mut self, cconv: &CallingConvention)

Remove all knowledge about the contents of callee-saved registers from the state.

pub fn specialize_by_expression_result(
&mut self,
expression: &Expression,
result: Data
) -> Result<(), Error>

Try to restrict the input variables of expression on self so that expression only evaluates to values represented by the given result.

If expression cannot evaluate to any value represented by self, return an error.

This function may restrict to upper bounds of possible values if the restriction cannot be made exact, i.e. after calling this function the state may still contain values for which expression does not evaluate to values represented by result.

impl State

pub fn to_json_compact(&self) -> Value

Get a more compact json-representation of the state. Intended for pretty printing, not useable for serialization/deserialization.