Module cwe_checker_lib::checkers

The implemented CWE checks. See their module descriptions for detailed information about each check.

Currently the Memory check is not contained in this module but directly incorporated into the pointer_inference module. See there for detailed information about this check.

Modules

cwe_78

This module implements a check for CWE-78: Improper Neutralization of Special Elements used in an OS Command (‘OS Command Injection’).

cwe_119

This module implements a check for CWE-119: Buffer Overflow and its variants CWE-125: Out-of-bounds Read and CWE-787: Out-of-bounds Write.

cwe_134

This module implements a check for CWE-134: Use of Externally-Controlled Format String.

cwe_190

This module implements a check for CWE-190: Integer overflow or wraparound.

cwe_215

This module implements a check for CWE-215: Information Exposure Through Debug Information.

cwe_243

This module implements a check for CWE-243: Creation of chroot Jail Without Changing Working Directory.

cwe_332

This module implements a check for CWE-332: Insufficient Entropy in PRNG.

cwe_337

This module implements a check for CWE-337: Predictable Seed in Pseudo-Random Number Generator (PRNG)

cwe_367

This module implements a check for CWE-367: Time-of-check Time-of-use (TOCTOU) Race Condition.

cwe_416

This module implements a check for CWE-415: Double Free and CWE-416: Use After Free.

cwe_426

This module implements a check for CWE-426: Untrusted Search Path.

cwe_467

This module implements a check for CWE-467: Use of sizeof() on a Pointer Type.

cwe_476

This module implements a check for CWE-476: NULL Pointer Dereference.

cwe_560

This module implements a check for CWE-560: Use of umask() with chmod-style Argument.

cwe_676

This module implements a check for CWE-676: Use of Potentially Dangerous Function.

cwe_782

This module implements a check for CWE-782: Exposed IOCTL with Insufficient Access Control.

cwe_789

This module implements a check for CWE-789: Memory Allocation with Excessive Size Value.