Struct cwe_checker_lib::intermediate_representation::ByteSize

pub struct ByteSize(_);

An unsigned number of bytes.

Used to represent sizes of values in registers or in memory. Can also be used for other byte-valued numbers, like offsets, as long as the number is guaranteed to be non-negative.

Implementations

impl ByteSize

pub fn new(value: u64) -> ByteSize

Create a new ByteSize object

pub fn as_bit_length(self) -> usize

Convert to the equivalent size in bits (by multiplying with 8).