Module cwe_checker_lib::pcode

Types to describe Ghidra P-Code and functions to translate it to the internally used intermediate representation.

The types in this module are not an exact representation of P-Code, as some preprocessing is already done in the P-Code-Extractor plugin.

The contents of this module are only used for the initial translation of P-Code into the internally used IR. For everything else the intermediate_representation should be used directly.

Structs

Arg
An argument (parameter or return value) of an extern symbol.

Blk
A basic block.

Call
A call instruction.

CallingConvention
A struct describing a calling convention.

Def
An assignment instruction, assigning the result of an expression to a varnode.

Expression
A P-Code expression.

ExternSymbol
An extern symbol, i.e. a function not contained in the binary but loaded from a shared library.

Jmp
A jump instruction.

Program
The program struct containing all information about the binary except for CPU-architecture-related information.

Project
The project struct describing all known information about the binary.

RegisterProperties
Properties of a register with respect to its base register.

Sub
A subfunction.

Variable
A variable representing a varnode in Ghidra P-Code

Enums

ArgIntent
The intent (input or output) of a function argument.

ExpressionType
Expression Opcodes as parsed from Ghidra

JmpType
A jump type mnemonic.

Label
A jump label for distinguishing between direct and indirect jumps.