Module cwe_checker_lib::utils

This module contains various utility modules and helper functions.

Modules

arguments

Handles argument detection by parsing format string arguments during a function call. (e.g. sprintf)

binary

Utility structs and functions which directly parse the binary file.

ghidra

Utility functions for executing Ghidra and extracting P-Code from the output.

graph_utils

Helper functions for common tasks utilizing the control flow graph of the binary.

log

Structs and functions for generating log messages and CWE warnings.

symbol_utils

Helper functions for common tasks utilizing extern symbols, e.g. searching for calls to a specific extern symbol.

Functions

get_binary_base_address

Get the base address for the image of a binary when loaded into memory.

get_ghidra_plugin_path

Get the folder path to a Ghidra plugin bundled with the cwe_checker.

read_config_file

Get the contents of a configuration file.