Function cwe_checker_lib::utils::graph_utils::is_sink_call_reachable_from_source_call

pub fn is_sink_call_reachable_from_source_call(
    graph: &Graph<'_>,
    source_node: NodeIndex,
    source_symbol: &Tid,
    sink_symbol: &Tid
) -> Option<Tid>

Check whether a call to the sink_symbol is reachable from the given source_node through a path of intraprocedural edges in the control flow graph.

A simple depth-first-search on the graph is used to find such a path. We do not search past subsequent calls to the source_symbol since we assume that sink calls after that belong to the new call to the source symbol and not the original one.

If a sink is found, the Tid of the jump term calling the sink is returned.